requirements.txt
attrs==24.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
logging==0.4.9.6
MarkupSafe==2.1.5
motor==3.5.1
MouseInfo==0.1.3
multidict==6.0.5
pyaes==1.6.1
pydub==0.25.1
pymongo==4.8.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
Pyrogram==2.0.106
PyScreeze==0.1.30
PySocks==1.7.1
python3-xlib==0.15
PyYAML==6.0.2
range==1.0.0
requests==2.32.3
TgCrypto==1.2.5
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4